fix: move collect_field_file module to utils
